Split textStyle plugin into one plugin per CSS property
