+++ defaultConfig.stub.js
@@ -48,7 +48,10 @@ module.exports = {
     textAlign: ['responsive'],
     textColor: ['responsive', 'hover', 'focus'],
     fontSize: ['responsive'],
-    textStyle: ['responsive', 'hover', 'focus'],
+    fontStyle: ['responsive', 'hover', 'focus'],
+    textTransform: ['responsive', 'hover', 'focus'],
+    textDecoration: ['responsive', 'hover', 'focus'],
+    fontSmoothing: ['responsive', 'hover', 'focus'],
     letterSpacing: ['responsive'],
     userSelect: ['responsive'],
     verticalAlign: ['responsive'],
