+++ plugins/textDecoration.js
@@ -1,0 +1,1 @@
+module.exports = require('../lib/plugins/textDecoration').default
