--- plugins/textStyle.js
@@ -1,1 +1,0 @@
-module.exports = require('../lib/plugins/textStyle').default
