+++ src/plugins/fontStyle.js
@@ -1,0 +1,11 @@
+export default function({ variants }) {
+  return function({ addUtilities }) {
+    addUtilities(
+      {
+        '.italic': { 'font-style': 'italic' },
+        '.roman': { 'font-style': 'normal' },
+      },
+      variants
+    )
+  }
+}
